requirements.txt
pysocks
requests
grequests
pandas
elasticsearch==7.1.0
mysql-connector-python==8.0.19
